requirements.txt
alembic==1.13.1
fastapi==0.110.1
httpx==0.27.0
networkx==3.2.1
psycopg2-binary==2.9.9
pydantic==2.6.4
pytest==8.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.29
uvicorn==0.29.0
